Updated code to allow it to run using xCode 7 (clang compiler)

+++ src/MapEditor/NewMapWindow.cpp
@@ -336,7 +336,7 @@ void NewMapWindow::onMapPropertiesChanged() {
 
     if(pMinimap != NULL) {
 
-        SDL_Rect dest = { 1, 1, pMinimap->w, pMinimap->h};
+        SDL_Rect dest = { 1, 1, static_cast<Uint16>(pMinimap->w), static_cast<Uint16>(pMinimap->h)};
         SDL_BlitSurface(pMinimap, NULL, pMapSurface, &dest);
         SDL_FreeSurface(pMinimap);
