+++ src/main.cpp
@@ -247,7 +247,7 @@ void printMissingFilesToScreen() {
     instruction += "\n\nPress ESC to exit.";
 
     SDL_Surface* pSurface = pFontManager->createSurfaceWithMultilineText(instruction, COLOR_BLACK, FONT_STD12);
-    SDL_Rect dest = { 30, 30, pSurface->w, pSurface->h };
+    SDL_Rect dest = { 30, 30, static_cast<Uint16>(pSurface->w), static_cast<Uint16>(pSurface->h) };
     SDL_BlitSurface(pSurface, NULL, screen, &dest);
     SDL_FreeSurface(pSurface);
